requirements.txt
streamlit==1.28.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-chroma==0.1.0
python-dotenv==1.0.0
chromadb==0.4.18
numpy==1.24.3
pdfplumber==0.9.0